requirements.txt
aider-chat==0.45.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-postgres-extensions==0.9.3
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
filelock==3.15.4
flake8==7.1.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grep-ast==0.3.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
litellm==1.41.6
mailtrap==2.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.10
openapi-codec==1.3.2
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pypandoc==1.13
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
sqlparse==0.4.4
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
wcwidth==0.2.13
whitenoise==6.7.0
yarl==1.9.4
zipp==3.19.2
